requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
filelock==3.18.0
Flask==3.1.0
fsspec==2025.3.2
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
pillow==11.2.1
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.13.2
Werkzeug==3.1.3
zipp==3.21.0
segment-anything==1.0
git+https://github.com/facebookresearch/segment-anything.git
pytest==8.0.0
albumentations==1.4.2